proceed update msg with kafka; auth

+++ src/main/java/com/rr/sociable/service/GroupService.java
@@ -5,7 +5,6 @@ import com.rr.sociable.entity.Group;
 import com.rr.sociable.entity.Message;
 import com.rr.sociable.entity.User;
 import com.rr.sociable.exception.NotFoundException;
-import com.rr.sociable.exception.UserAlreadyInGroupException;
 import com.rr.sociable.repo.GroupRepository;
 import com.rr.sociable.repo.MessageRepository;
 import com.rr.sociable.repo.UserRepository;
@@ -16,6 +15,7 @@ import org.springframework.data.domain.Pageable;
 
 
 import java.util.List;
+import java.util.Optional;
 import java.util.Set;
 
 @Service
@@ -34,36 +34,66 @@ public class GroupService {
         return groupRepository.findAll();
     }
 
-    public Group findById(Long id) {
-         return groupRepository.findById(id)
-                .orElseThrow(() -> new NotFoundException(String.format("User with id %d not found", id)));
+    public Optional<Group> findById(Long id) {
+         return groupRepository.findById(id);
     }
 
     @Transactional
-    public void addUserToGroup(Long groupId, Long userId) {
-        Group group = groupRepository.findById(groupId)
-                .orElseThrow(() -> new NotFoundException("Group not found"));
-        User user = userRepository.findById(userId).orElseThrow(() -> new NotFoundException("User not found"));
+    public boolean addUserToGroup(Long groupId, Long userId) {
+        Optional<Group> g = groupRepository.findById(groupId);
+        Optional<User> u = userRepository.findById(userId);
+        if (g.isEmpty() || u.isEmpty()) {
+            return false;
+        }
+
+        Group group = g.get();
 
         if(group.getMemberIds().contains(userId)){
-           throw new UserAlreadyInGroupException("User already in group");
+            return false;
         }
 
         group.getMemberIds().add(userId);
         groupRepository.save(group);
 
+        User user = u.get();
+
         user.getGroupsIds().add(groupId);
         userRepository.save(user);
+        return true;
     }
 
     @Transactional
-    public Group save(GroupDto groupDto) {
+    public Optional<Group> save(GroupDto groupDto) {
         Group group = new Group(groupDto);
-        return groupRepository.save(group);
+        group.setMemberCount(1);
+        Optional<User> u = userRepository.findById(groupDto.getCreatorId());
+
+        if(u.isEmpty()) return Optional.empty();
+
+        User user = u.get();
+
+        group.getMemberIds().add(user.getId());
+
+        Group saved = groupRepository.save(group);
+        user.getGroupsIds().add(saved.getId());
+
+        userRepository.save(user);
+
+        return Optional.of(saved);
     }
 
     @Transactional
     public void delete(Long id) {
+        Optional<Group> g = groupRepository.findById(id);
+        if(g.isEmpty()) return;
+
+        Group group = g.get();
+        List<User> users = userRepository.findAllById(group.getMemberIds());
+        users.forEach(u -> u.getGroupsIds().remove(group.getId()));
+        userRepository.saveAll(users);
+
+        messageRepository.deleteMessageByGroupId(id);
+
         groupRepository.deleteById(id);
     }
 
@@ -83,4 +113,11 @@ public class GroupService {
         return messageRepository.findAllByGroupId(pageable, groupId);
     }
 
+    public Optional<Group> update(Long id, GroupDto group) {
+        Optional<Group> g = groupRepository.findById(id);
+        if (g.isEmpty()) return Optional.empty();
+        Group findedGroup = g.get();
+        findedGroup.setName(group.getName());
+        return Optional.of(groupRepository.save(findedGroup));
+    }
 }
